# Fix data race in BundleRegistry::Install()