--- a/framework/src/bundle/BundleRegistry.cpp
+++ b/framework/src/bundle/BundleRegistry.cpp
@@ -44,12 +44,14 @@
 namespace cppmicroservices {
 
 // Helper class to ensure RAII for InitialBundleMap in case where Install0 throws
-class InitialBundleMapCleanup {
+class InitialBundleMapCleanup
+{
 public:
-  InitialBundleMapCleanup(std::function<void()> cleanupFcn) : _cleanupFcn(std::move(cleanupFcn)) {}
-  ~InitialBundleMapCleanup() {
-	_cleanupFcn();
-  }
+  InitialBundleMapCleanup(std::function<void()> cleanupFcn)
+    : _cleanupFcn(std::move(cleanupFcn))
+  {}
+  ~InitialBundleMapCleanup() { _cleanupFcn(); }
+
 private:
   std::function<void()> _cleanupFcn;
 };
@@ -93,18 +95,22 @@ void BundleRegistry::DecrementInitialBundleMapRef(
 /*
   This function populates the res and alreadyInstalled vectors with the
   appropriate entries so that they can be used by the Install0 call. This was
-  extracted from Install() for convenience.
+  extracted from Install() for convenience. We lock the 'bundles' object to
+  prevent any installs from writing to the map while this operation is occurring.
 */
 void BundleRegistry::GetAlreadyInstalledBundlesAtLocation(
   std::pair<BundleMap::iterator, BundleMap::iterator> foundBundles,
   std::vector<Bundle>& resultingBundles,
   std::vector<std::shared_ptr<BundlePrivate>>& alreadyInstalled)
 {
+  auto l = bundles.Lock();
+  US_UNUSED(l);
   while (foundBundles.first != foundBundles.second) {
     auto installedBundlePrivate = foundBundles.first->second;
     alreadyInstalled.push_back(installedBundlePrivate);
     auto actualBundle = coreCtx->bundleHooks.FilterBundle(
-      MakeBundleContext(installedBundlePrivate->bundleContext.Load()), MakeBundle(installedBundlePrivate));
+      MakeBundleContext(installedBundlePrivate->bundleContext.Load()),
+      MakeBundle(installedBundlePrivate));
     if (actualBundle) {
       resultingBundles.push_back(actualBundle);
     }
@@ -125,7 +131,8 @@ std::vector<Bundle> BundleRegistry::Install(const std::string& location,
   US_UNUSED(l);
 
   // Search the multimap for the current bundle location
-  auto bundlesAtLocationRange = (bundles.Lock(), bundles.v.equal_range(location));
+  auto bundlesAtLocationRange =
+    (bundles.Lock(), bundles.v.equal_range(location));
 
   /*
     If the bundle is already installed, then execute the regular
@@ -133,8 +140,8 @@ std::vector<Bundle> BundleRegistry::Install(const std::string& location,
 
     If the bundle isn't installed, then one of two things can happen: 1) either
     the current thread is the first thread trying to install this bundle or 2) the
-    current thread is trying to install a bundle that is not installed which another
-    thread is currently installing.
+    current thread is trying to install a bundle that is not installed but another
+    thread is currently installing that bundle.
 
     If 1): Create an entry in the initialBundleInstallMap which keeps track of whether
       or not a given bundle is being installed for the first time. After creating this
@@ -151,16 +158,17 @@ std::vector<Bundle> BundleRegistry::Install(const std::string& location,
   */
   if (bundlesAtLocationRange.first != bundlesAtLocationRange.second) {
     l.UnLock();
-
     std::vector<Bundle> resultingBundles;
     std::vector<std::shared_ptr<BundlePrivate>> alreadyInstalled;
     // Populate the res and alreadyInstalled vectors with the appropriate data
     // based on what bundles are already installed
-    GetAlreadyInstalledBundlesAtLocation(bundlesAtLocationRange, resultingBundles, alreadyInstalled);
+    GetAlreadyInstalledBundlesAtLocation(
+      bundlesAtLocationRange, resultingBundles, alreadyInstalled);
 
     // Perform the install
     auto newBundles = Install0(location, alreadyInstalled, caller);
-    resultingBundles.insert(resultingBundles.end(), newBundles.begin(), newBundles.end());
+    resultingBundles.insert(
+      resultingBundles.end(), newBundles.begin(), newBundles.end());
     if (resultingBundles.empty()) {
       throw std::runtime_error("All bundles rejected by a bundle hook");
     } else {
@@ -173,7 +181,7 @@ std::vector<Bundle> BundleRegistry::Install(const std::string& location,
 
     /*
       If no other thread is installing the desired bundle, create a map entry,
-      install the bundle, and notify all other threads wanting to install this bundle
+      install the bundle, and notify all other threads waiting to install this bundle
       that it is safe to do so. If the current thread is not the installing thread, then
       it will increment the reference count in the initialBundleInstallMap so that the map
       entry isn't prematurely deleted, wait to be notified that the install thread is done,
@@ -181,26 +189,29 @@ std::vector<Bundle> BundleRegistry::Install(const std::string& location,
     */
     if (alreadyInstallingIterator == initialBundleInstallMap.end()) {
       // Insert entry into the initialBundleInstallMap to prevent other threads from
-      // trying to install this uninstalled bundle at the same time
+      // trying to install this bundle simultaneously at the same time
       auto pairToInsert = std::make_pair(uint32_t(1), WaitCondition{});
       initialBundleInstallMap.insert(
         std::make_pair(location, std::move(pairToInsert)));
       l.UnLock();
 
       std::vector<Bundle> installedBundles;
-      
       {
         // create instance of clean-up object to ensure RAII
-        InitialBundleMapCleanup cleanup([this, &l, &location](){
+        InitialBundleMapCleanup cleanup([this, &l, &location]() {
           {
+            l.Lock();
+            auto& p = initialBundleInstallMap[location];
+
             // Notify all waiting threads that it is safe to install the bundle
-            std::lock_guard<std::mutex> lock(*(initialBundleInstallMap[location].second.m));
-            initialBundleInstallMap[location].second.waitFlag = false;
-            initialBundleInstallMap[location].second.cv->notify_all();
+            std::lock_guard<std::mutex> lock(*(p.second.m));
+            p.second.waitFlag = false;
+            p.second.cv->notify_all();
+            l.UnLock();
           }
           DecrementInitialBundleMapRef(l, location);
         });
-          
+
         // Perform the install
         installedBundles = Install0(location, {}, caller);
       }
@@ -208,37 +219,39 @@ std::vector<Bundle> BundleRegistry::Install(const std::string& location,
       return installedBundles;
     } else {
       initialBundleInstallMap[location].first++;
-      
       {
         // Wait for the install thread to notify this thread that it is safe
         // to install the current bundle
-        std::unique_lock<std::mutex> lock(
-          *(initialBundleInstallMap[location].second.m));
+        auto& p = initialBundleInstallMap[location];
+        l.UnLock();
+        std::unique_lock<std::mutex> lock(*(p.second.m));
 
         // This while loop exists to prevent a known race condition. If the installing thread notifies before
         // another thread trying to install the same bundle reaches this wait, it will wait indefinitely. To
-        // fix this, a wait_for is used intead (which utilizes a timeout to avoid this race) and the wait statement
+        // fix this, a wait_for is used instead (which utilizes a timeout to avoid this race) and the wait statement
         // as a whole acts as the while statement's predicate; once the timeout is reached, wait_for exits and the
         // statement is re-evaluated since it would have returned false.
-        while (!initialBundleInstallMap[location].second.cv->wait_for(lock, 0.1ms, [&location, this] {
-          return !initialBundleInstallMap[location].second.waitFlag;
-          }));
+        while (!p.second.cv->wait_for(
+          lock, 0.01ms, [&p] { return !p.second.waitFlag; }))
+          ;
       }
-      
+
+      l.Lock();
       // Re-acquire the range because while this thread was waiting, the installing
       // thread made a modification to bundles.v
-      bundlesAtLocationRange = (bundles.Lock(), bundles.v.equal_range(location));
+      bundlesAtLocationRange =
+        (bundles.Lock(), bundles.v.equal_range(location));
       l.UnLock();
 
       std::vector<Bundle> resultingBundles;
       std::vector<std::shared_ptr<BundlePrivate>> alreadyInstalled;
-      GetAlreadyInstalledBundlesAtLocation(bundlesAtLocationRange, resultingBundles, alreadyInstalled);
+      GetAlreadyInstalledBundlesAtLocation(
+        bundlesAtLocationRange, resultingBundles, alreadyInstalled);
 
       std::vector<Bundle> newBundles;
-      
       {
         // create instance of clean-up object to ensure RAII
-        InitialBundleMapCleanup cleanup([this, &l, &location](){
+        InitialBundleMapCleanup cleanup([this, &l, &location]() {
           DecrementInitialBundleMapRef(l, location);
         });
 
@@ -246,7 +259,8 @@ std::vector<Bundle> BundleRegistry::Install(const std::string& location,
         newBundles = Install0(location, alreadyInstalled, caller);
       }
 
-      resultingBundles.insert(resultingBundles.end(), newBundles.begin(), newBundles.end());
+      resultingBundles.insert(
+        resultingBundles.end(), newBundles.begin(), newBundles.end());
       if (resultingBundles.empty()) {
         throw std::runtime_error("All bundles rejected by a bundle hook");
       } else {
@@ -415,7 +429,7 @@ void BundleRegistry::Load()
       ba->SetAutostartSetting(-1); // Do not start on launch
       std::cerr << "Failed to load bundle " << util::ToString(ba->GetBundleId())
                 << " (" + ba->GetBundleLocation() + ") uninstalled it!"
-                << " (execption: "
+                << " (exception: "
                 << util::GetExceptionStr(std::current_exception()) << ")"
                 << std::endl;
     }