This PR addresses the data race currently in BundleRegistry::Install() which causes sporadic crashes (segfaults) when a large number of bundles are being installed at the same time. Performance benefits are exhibited in this change and significantly reduce the amount of time it takes to install all bundles when 2-4 threads are used. There could exist a larger configuration of threads which is faster, but with the input size used to test (~2048 bundles), it appears to be the best.

All benchmarking was performed on a Windows 10, 64bit machine (i7-4790k @ 4.0GHz) with 2048 bundles.

Performance (note: CPU numbers don't mean much, most of the runs during the benchmark had 0 for CPU):
| Benchmark                                                                     | Time (before PR #411, ns) | Time (fix-concurrent-bundle-install, ns) | CPU (before PR #411, ns) | CPU (fix-concurrent-bundle-install, ns) | Iterations |
|-------------------------------------------------------------------------------|--------------------------:|-----------------------------------------:|-------------------------:|----------------------------------------:|-----------:|
| BundleInstallFixture/ConcurrentBundleInstall1Thread/manual_time_mean          |                  12072893 |                                 10665267 |                    12850 |                                   10615 |         64 |
| BundleInstallFixture/ConcurrentBundleInstall2Threads/manual_time_mean         |                   9031489 |                                  4981662 |                    15751 |                                   33203 |         64 |
| BundleInstallFixture/ConcurrentBundleInstall4Threads/manual_time_mean         |                  18439921 |                                  5841105 |                    32552 |                                   58235 |         64 |
| BundleInstallFixture/ConcurrentBundleInstallMaxThreads/manual_time_mean       |                  26885067 |                                 13754601 |                    24414 |                                   44389 |         64 |
| BundleInstallFixture/ConcurrentBundleInstall1ThreadPerBundle/manual_time_mean |                  61213356 |                                 17298247 |                  3662109 |                                 9580870 |         64 |

The existing test suite is incapable of detecting this issue due to the number of bundles being installed is minuscule in comparison to the number of bundles required to see this data race.

To reproduce the issue, a large number of bundles (~2048) were generated and a small program was written to install all of them on many threads at the same time (mimicking the BundleRegistryConcurrencyTest test that we currently have except on a larger scale). When TSAN was turned on, warnings started to appear regarding a data race.

Having a test point to capture this behavior is out of the scope for this PR but I believe it to be valuable when working with this code in the future due to the nature of the issue.